KIWI-2169 Alarm for PO failure emails
